requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic[email]
python-dotenv
